requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
bcrypt==3.2.0
beautifulsoup4==4.9.3
bootstrap4==0.1.0
certifi==2021.5.30
cffi==1.14.3
chardet==4.0.0
cloudinary==1.26.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.5
Django==3.1.1
django-betterforms==1.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==14.1.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2
django-email-verification==0.1.0
django-extra-views==0.14.0
django-misaka==0.2.1
Django-Verify-Email==1.0.7
djangorestframework==3.12.4
Faker==4.1.3
gunicorn==20.1.0
houdini.py==0.1.0
idna==2.10
misaka==2.1.1
Naked==0.1.31
paytmchecksum==1.7.0
Pillow==7.2.0
psycopg2-binary==2.9.1
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.7.3
PyJWT==1.7.1
pyotp==2.6.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.1
PyYAML==5.4.1
requests==2.25.1
shellescape==3.8.1
six==1.15.0
soupsieve==2.1
sqlparse==0.3.1
text-unidecode==1.3
twilio==6.60.0
urllib3==1.26.6
whitenoise==5.2.0
